moved competition files to their own package and made made some programs for a demo robot
