+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Competition/R1FullAuto.java
@@ -1,5 +1,5 @@
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Competition;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -19,8 +19,8 @@ import org.openftc.easyopencv.OpenCvInternalCamera;
 import org.openftc.easyopencv.OpenCvPipeline;
 
 
-@Autonomous
-public class R1ScanParkDelay extends LinearOpMode {
+@Autonomous(name="Red Full 1")
+public class R1FullAuto extends LinearOpMode {
 
     private static final long SLEEP_10 = 10;
     private static final long SLEEP_25 = 25;
@@ -77,7 +77,7 @@ public class R1ScanParkDelay extends LinearOpMode {
 
 
         waitForStart();
-        sleep(globalvar.delaytime);
+
 
         double FORWARD_SPEED = 0.5;
 
@@ -91,19 +91,16 @@ public class R1ScanParkDelay extends LinearOpMode {
 
 
 
-
-
-            encoderDriveWithoutTime(.5,1,1,1,1);
             robot.Slider.setPower(1);
-            sleep(600);
-            robot.Slider.setPower(0.2);
+            sleep(400);
+            robot.Slider.setPower(0);
             sleep(200);
-            double moveLength = 33.5;
+            double moveLength = 28;
             int position = 1;
 
-            //spinAndComeBack();
+            spinAndComeBack();
             //stop();
-            robotMoveToShippingElement(13);
+            robotMoveToShippingElement(6.5);
 
 
             boolean positionShippingElement = scanShippingElement();
@@ -113,12 +110,12 @@ public class R1ScanParkDelay extends LinearOpMode {
             if (positionShippingElement){
                 telemetry.addData("The position is",position);
                 telemetry.update();
-                moveToShippingHub(moveLength, position);
+                moveToShippingHub(moveLength + 5, position);
             }
 
             else {
 
-                moveToSecondDuck(8);
+                moveToSecondDuck(9.5);
                 positionShippingElement = scanShippingElement();
                 telemetry.addData("Is the shipping element present",positionShippingElement);
                 telemetry.update();
@@ -128,14 +125,14 @@ public class R1ScanParkDelay extends LinearOpMode {
                     telemetry.addData("The position is",position);
                     telemetry.update();
 
-                    moveToShippingHub(moveLength - 8, position);
+                    moveToShippingHub(moveLength - 4.5, position);
                 }
                 else {
                     position = 3;
                     telemetry.addData("The position is",position);
                     telemetry.update();
 
-                    moveToShippingHub(moveLength - 6, position);
+                    moveToShippingHub(moveLength - 1, position);
                 }
             }
 
@@ -164,19 +161,35 @@ public class R1ScanParkDelay extends LinearOpMode {
 
     private void spinAndComeBack() {
 
-        encoderDriveWithoutTime(.5,4,-4,4,-4);
+        //encoderDriveWithoutTime(.5,4,-4,4,-4);
+        //sleep(100);
+        encoderDriveWithoutTime(.5,-1,-1,-1,-1);
+        sleep(50);
+
+        encoderDriveWithoutTime(.5,-20,20,-20,20);
+        sleep(50);
+        encoderDriveWithoutTime(.1,-2.5,2.5,-2.5,2.5);
         sleep(100);
-        encoderDriveWithoutTime(.3,26,26,26,26);
-        sleep(100);
-        spinLeft(4,.4);
-        sleep(100);
-        encoderDriveWithoutTime(.5,-1,1,1,-1);
+        //encoderDriveWithoutTime(.3,50,50,50,50);
+        //sleep(100);
+        spinRight(2.5,.4);
+        sleep(50);
+       //strafes left
+        encoderDriveWithoutTime(.5,11.5,-11.5,11.5,-11.5);
+        encoderDriveWithoutTime(.7,-6,-6,-6,-6);
+        turn90LeftMore();
+        turn90LeftMore();
+        sleep(50);
+        encoderDriveWithoutTime(0.3,-2.5,2.5,2.5,-2.5);
+        sleep(50);
+
+        /*encoderDriveWithoutTime(.5,-1,1,1,-1);
         sleep(100);
         encoderDriveWithoutTime(.5,6,-6,6,-6);
         sleep(100);
         encoderDriveWithoutTime(.5,-19,-19,-19,-19);
         sleep(100);
-        turn90Left();
+        turn90Left(); */
     }
 
     private void turn90Left() {
@@ -198,7 +211,7 @@ public class R1ScanParkDelay extends LinearOpMode {
         //encoderDriveWithoutTime(0.5,4,4,4,4);
         //go forward
         //encoderDriveWithoutTime(0.5, 48,48,48,48);
-        encoderDriveWithoutTime(.8,5,5,5,5);
+        encoderDriveWithoutTime(.8,8,8,8,8);
         sleep(100);
         turn90Left();
         telemetry.addData("Before final move to warehouse","");
@@ -213,46 +226,50 @@ public class R1ScanParkDelay extends LinearOpMode {
             case 1:
                 //adjust arm position and drop cube
                 double len = 2;
+                encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
+                encoderDriveWithoutTime(.4,-3,-3,-3,-3);
 
-                robot.Sweeper.setPosition(0.4);
-                sleep(50);
-                len = 2;
-                encoderDriveWithoutTime(.5,-len,-len,-len,-len);
-                sleep(100);
+
+                robot.Intake1.setPower(0.8);
+                robot.Intake2.setPower(-0.8);
+                sleep(2000);
+                robot.Intake2.setPower(0);
+                robot.Intake1.setPower(0);
+
+
                 break;
-
-
-
-
             case 2:
                 robot.Slider.setPower(0.7);
-                sleep(900);
-                robot.Slider.setPower(0.2);
-
-
-                robot.Sweeper.setPosition(0.4);
-                sleep(50);
-                len = 2;
-                encoderDriveWithoutTime(.5,-len,-len,-len,-len);
+                sleep(450);
+                robot.Slider.setPower(0.1);
                 sleep(100);
+                encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
+
+                encoderDriveWithoutTime(.4,-3,-3,-3,-3);
+                robot.Intake1.setPower(0.8);
+                robot.Intake2.setPower(-0.8);
+                sleep(2000);
+                robot.Intake2.setPower(0);
+                robot.Intake1.setPower(0);
+
                 //adjust arm position and drop cube
                 break;
             default:
-
-
                 robot.Slider.setPower(1);
-                sleep(1300);
+                sleep(800);
                 robot.Slider.setPower(0.1);
-
-                robot.Sweeper.setPosition(0);
-                sleep(50);
-                len = 2;
-                encoderDriveWithoutTime(.5,-len,-len,-len,-len);
                 sleep(100);
-                //adjust arm position and drop cube
+                encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
+                encoderDriveWithoutTime(.4,-3,-3,-3,-3);
+                robot.Intake1.setPower(0.8);
+                robot.Intake2.setPower(-0.8);
+                sleep(2000);
+                robot.Intake2.setPower(0);
+                robot.Intake1.setPower(0);
+
+
+                //adjust arm position and drop c4ube
                 break;
-
-
 
         }
 
@@ -266,16 +283,16 @@ public class R1ScanParkDelay extends LinearOpMode {
 
     private void moveToShippingHub(double moveLength, int position) {
 
-        encoderDriveWithoutTime(0.5, -moveLength, moveLength, -moveLength, moveLength);
+        encoderDriveWithoutTime(0.5, -moveLength - 2, moveLength + 2, -moveLength - 2, moveLength + 2);
         sleep(100);
         turn90LeftMore();
         telemetry.addData("Its working", "");
         telemetry.update();
         turn90LeftMore();
         sleep(100);
-        //robot.Sweeper.setPosition(0);
         encoderDriveWithoutTime(0.5,4,-4,4,-4);
-        encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
+        //robot.Sweeper.setPosition(0);
+
 
         //encoderDriveWithoutTime(.3, -6,6,-6,6);
         //placeFreightCorrectLocation(position);
@@ -286,7 +303,7 @@ public class R1ScanParkDelay extends LinearOpMode {
         sleep(1000);
         int avg1 = pipeline.getAnalysis();
 
-        final int THRESHOLD = 150;//150;
+        final int THRESHOLD = 155;//150;
         telemetry.addData("Scan Threshhold:",avg1);
         telemetry.update();
         sleep(1000);
@@ -335,6 +352,7 @@ public class R1ScanParkDelay extends LinearOpMode {
         telemetry.addData("Going toward Shipping element", tmpmoveLength);
         telemetry.update();
         encoderDriveWithoutTime(0.3, tmpmoveLength, tmpmoveLength, tmpmoveLength, tmpmoveLength);
+
         //encoderDriveWithTime(1,-1,-1,-1,-1,10);
     }
