+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Competition/R1FullAutoDelay.java
@@ -1,7 +1,5 @@
 
-package org.firstinspires.ftc.teamcode;
-
-import static org.firstinspires.ftc.teamcode.zanehardware.COUNTS_PER_INCH;
+package org.firstinspires.ftc.teamcode.Competition;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -21,8 +19,8 @@ import org.openftc.easyopencv.OpenCvInternalCamera;
 import org.openftc.easyopencv.OpenCvPipeline;
 
 
-@Autonomous
-public class B1SpinPark extends LinearOpMode {
+@Autonomous(name="Red Full 1 delay")
+public class R1FullAutoDelay extends LinearOpMode {
 
     private static final long SLEEP_10 = 10;
     private static final long SLEEP_25 = 25;
@@ -37,7 +35,7 @@ public class B1SpinPark extends LinearOpMode {
     static final double     COUNTS_PER_MOTOR_REV    = 537.6;//356.3 ;    // eg: DC Motor Encoder
     static final double     DRIVE_GEAR_REDUCTION    = 1.0 ;     // This is < 1.0 if geared UP
     static final double     WHEEL_DIAMETER_INCHES   = 3.77953 ;     // For figuring circumference
-    //static final double     COUNTS_PER_INCH = 45;
+    static final double     COUNTS_PER_INCH = 45;
             //first hundred digits of pi fr more accuracy
 
 
@@ -75,11 +73,16 @@ public class B1SpinPark extends LinearOpMode {
 
         robot.init(hardwareMap);
 
-        waitForStart();
 
+
+
+        waitForStart();
+        sleep(globalvar.delaytime);
 
         double FORWARD_SPEED = 0.5;
 
+
+        //scanShippingElementTest();
 
 
         while (opModeIsActive()) {
@@ -87,66 +90,49 @@ public class B1SpinPark extends LinearOpMode {
 
 
 
-            encoderDriveWithoutTime(.5,4,-4,4,-4);
-            sleep(100);
-            encoderDriveWithoutTime(.3,26,26,26,26);
-            sleep(100);
-            spinLeft(5,.4);
-            sleep(100);
-            encoderDriveWithoutTime(.5,3,-3,3,-3);
-            sleep(100);
-            encoderDriveWithoutTime(.8,-107,-107,-107,-107);
-            stop();
-            //encoderDriveWithoutTimeFrontLeft(1,5,5,5,5);
-            //encoderDriveWithoutTimeFrontRight(1,5,5,5,5);
-            //stop();,
-            encoderDriveWithoutTime(.3,5,5,5,5);
-            sleep(1000);
-            encoderDriveWithoutTime(.3,-5,-5,-5,-5);
-            sleep(1000);
-            encoderDriveWithoutTime(.3,5,-5,5,-5);
-            sleep(1000);
-            encoderDriveWithoutTime(.3,-5,5,-5,5);
-            stop();
-            /*sleep(100);
-            encoderDriveWithTimeForward(2.7,.5);
-            sleep(100);
-            spinRight(7, .3);
-            sleep(100);
-            encoderDriveWithTimeLeft(.4,.5);
-            sleep(100);
-            encoderDriveWithTimeBackward(6,1);
-            stop();*/
 
+            robot.Slider.setPower(1);
+            sleep(400);
+            robot.Slider.setPower(0);
+            sleep(200);
             double moveLength = 28;
             int position = 1;
-            robotMoveToShippingElement(1);
-            sleep(5000);
+
+            spinAndComeBack();
+            //stop();
+            robotMoveToShippingElement(6.5);
+
 
             boolean positionShippingElement = scanShippingElement();
             telemetry.addData("Is the shipping element present",positionShippingElement);
             telemetry.update();
-            sleep(1000);
+            sleep(100);
             if (positionShippingElement){
-                moveToShippingHub(moveLength, position);
+                telemetry.addData("The position is",position);
+                telemetry.update();
+                moveToShippingHub(moveLength + 5, position);
             }
 
             else {
 
-                moveToSecondDuck(6);
+                moveToSecondDuck(9.5);
                 positionShippingElement = scanShippingElement();
                 telemetry.addData("Is the shipping element present",positionShippingElement);
                 telemetry.update();
-                sleep(1000);
+                sleep(100);
                 if (positionShippingElement){
                     position = 2;
+                    telemetry.addData("The position is",position);
+                    telemetry.update();
 
-                    moveToShippingHub(moveLength - 6, position);
+                    moveToShippingHub(moveLength - 4.5, position);
                 }
                 else {
                     position = 3;
+                    telemetry.addData("The position is",position);
+                    telemetry.update();
 
-                    moveToShippingHub(moveLength - 6, position);
+                    moveToShippingHub(moveLength - 1, position);
                 }
             }
 
@@ -173,11 +159,65 @@ public class B1SpinPark extends LinearOpMode {
         }
     }
 
+    private void spinAndComeBack() {
+
+        //encoderDriveWithoutTime(.5,4,-4,4,-4);
+        //sleep(100);
+        encoderDriveWithoutTime(.5,-1,-1,-1,-1);
+        sleep(50);
+
+        encoderDriveWithoutTime(.5,-20,20,-20,20);
+        sleep(50);
+        encoderDriveWithoutTime(.1,-2.5,2.5,-2.5,2.5);
+        sleep(100);
+        //encoderDriveWithoutTime(.3,50,50,50,50);
+        //sleep(100);
+        spinRight(2.5,.4);
+        sleep(50);
+       //strafes left
+        encoderDriveWithoutTime(.5,11.5,-11.5,11.5,-11.5);
+        encoderDriveWithoutTime(.7,-6,-6,-6,-6);
+        turn90LeftMore();
+        turn90LeftMore();
+        sleep(50);
+        encoderDriveWithoutTime(0.3,-2.5,2.5,2.5,-2.5);
+        sleep(50);
+
+        /*encoderDriveWithoutTime(.5,-1,1,1,-1);
+        sleep(100);
+        encoderDriveWithoutTime(.5,6,-6,6,-6);
+        sleep(100);
+        encoderDriveWithoutTime(.5,-19,-19,-19,-19);
+        sleep(100);
+        turn90Left(); */
+    }
+
+    private void turn90Left() {
+
+        encoderDriveWithoutTime(.5,-19.5,19.5,19.5,-19.5);
+    }
+
+
+    private void turn90LeftMore() {
+
+        encoderDriveWithoutTime(.5,-20.5,20.5,20.5,-20.5);
+    }
+    private void turn90Right() {
+
+        encoderDriveWithoutTime(.5,18.5,-18.5,-18.5,18.5);
+    }
     private void moveToWarehouse() {
         //turn to the left for blue
-        encoderDriveWithoutTime(0.5,4,4,4,4);
+        //encoderDriveWithoutTime(0.5,4,4,4,4);
         //go forward
-        encoderDriveWithoutTime(0.5, 48,48,48,48);
+        //encoderDriveWithoutTime(0.5, 48,48,48,48);
+        encoderDriveWithoutTime(.8,8,8,8,8);
+        sleep(100);
+        turn90Left();
+        telemetry.addData("Before final move to warehouse","");
+        telemetry.update();
+        encoderDriveWithoutTime(.8,65,65,65,65);
+
     }
 
     private void placeFreightCorrectLocation(int position) {
@@ -185,12 +225,50 @@ public class B1SpinPark extends LinearOpMode {
         switch (position) {
             case 1:
                 //adjust arm position and drop cube
+                double len = 2;
+                encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
+                encoderDriveWithoutTime(.4,-3,-3,-3,-3);
+
+
+                robot.Intake1.setPower(0.8);
+                robot.Intake2.setPower(-0.8);
+                sleep(2000);
+                robot.Intake2.setPower(0);
+                robot.Intake1.setPower(0);
+
+
                 break;
             case 2:
+                robot.Slider.setPower(0.7);
+                sleep(450);
+                robot.Slider.setPower(0.1);
+                sleep(100);
+                encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
+
+                encoderDriveWithoutTime(.4,-3,-3,-3,-3);
+                robot.Intake1.setPower(0.8);
+                robot.Intake2.setPower(-0.8);
+                sleep(2000);
+                robot.Intake2.setPower(0);
+                robot.Intake1.setPower(0);
+
                 //adjust arm position and drop cube
                 break;
             default:
-                //adjust arm position and drop cube
+                robot.Slider.setPower(1);
+                sleep(800);
+                robot.Slider.setPower(0.1);
+                sleep(100);
+                encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
+                encoderDriveWithoutTime(.4,-3,-3,-3,-3);
+                robot.Intake1.setPower(0.8);
+                robot.Intake2.setPower(-0.8);
+                sleep(2000);
+                robot.Intake2.setPower(0);
+                robot.Intake1.setPower(0);
+
+
+                //adjust arm position and drop c4ube
                 break;
 
         }
@@ -199,15 +277,25 @@ public class B1SpinPark extends LinearOpMode {
 
     private void moveToSecondDuck(double moveLength) {
 
-        //encoderDriveWithoutTime(0.3, moveLength, moveLength, moveLength, moveLength);
-        encoderDriveWithTime(0.3,1,1,1,1,2);
+        encoderDriveWithoutTime(0.3, -moveLength, moveLength, -moveLength, moveLength);
+        //encoderDriveWithTime(0.3,1,1,1,1,2);
     }
 
     private void moveToShippingHub(double moveLength, int position) {
 
-        encoderDriveWithoutTime(0.5, -moveLength, -moveLength, -moveLength, -moveLength);
-        encoderDriveWithoutTime(.3, -6,6,-6,6);
-        placeFreightCorrectLocation(position);
+        encoderDriveWithoutTime(0.5, -moveLength - 2, moveLength + 2, -moveLength - 2, moveLength + 2);
+        sleep(100);
+        turn90LeftMore();
+        telemetry.addData("Its working", "");
+        telemetry.update();
+        turn90LeftMore();
+        sleep(100);
+        encoderDriveWithoutTime(0.5,4,-4,4,-4);
+        //robot.Sweeper.setPosition(0);
+
+
+        //encoderDriveWithoutTime(.3, -6,6,-6,6);
+        //placeFreightCorrectLocation(position);
     }
 
     private boolean scanShippingElement() {
@@ -215,9 +303,40 @@ public class B1SpinPark extends LinearOpMode {
         sleep(1000);
         int avg1 = pipeline.getAnalysis();
 
+        final int THRESHOLD = 155;//150;
+        telemetry.addData("Scan Threshhold:",avg1);
+        telemetry.update();
+        sleep(1000);
+
+
+        if(avg1 > THRESHOLD){
+            return true;
+        }else{
+            return false;
+        }
+
+
+    }
+    private boolean scanShippingElementTest() {
+
+        while (opModeIsActive()) {
+            sleep(100);
+            int avg1 = pipeline.getAnalysis();
+
+            final int THRESHOLD = 25;//150;
+            telemetry.addData("Scan Threshhold:",avg1);
+            telemetry.update();
+            sleep(2000);
+
+
+        }
+        sleep(100);
+        int avg1 = pipeline.getAnalysis();
+
         final int THRESHOLD = 25;//150;
         telemetry.addData("Scan Threshhold:",avg1);
         telemetry.update();
+        sleep(5000);
 
 
         if(avg1 > THRESHOLD){
@@ -232,8 +351,9 @@ public class B1SpinPark extends LinearOpMode {
     private void robotMoveToShippingElement(double tmpmoveLength) {
         telemetry.addData("Going toward Shipping element", tmpmoveLength);
         telemetry.update();
-        //encoderDriveWithoutTime(0.3, -tmpmoveLength, -tmpmoveLength, -tmpmoveLength, -tmpmoveLength);
-        encoderDriveWithTime(1,-1,-1,-1,-1,10);
+        encoderDriveWithoutTime(0.3, tmpmoveLength, tmpmoveLength, tmpmoveLength, tmpmoveLength);
+
+        //encoderDriveWithTime(1,-1,-1,-1,-1,10);
     }
 
 
@@ -316,155 +436,6 @@ public class B1SpinPark extends LinearOpMode {
             robot.Back_Left.setPower(0);
             robot.Back_Right.setPower(0);
             robot.Front_Left.setPower(0);
-            robot.Front_Right.setPower(0);
-
-            // Turn off RUN_TO_POSITION
-            robot.Back_Left.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            robot.Back_Right.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            robot.Front_Left.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            robot.Front_Right.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-            //  sleep(250);   // optional pause after each move
-
-        }
-
-    }
-    public void encoderDriveWithoutTimeFrontLeft ( double speed,
-                                          double Back_Left_Inches,
-                                          double Back_Right_Inches,
-                                          double Front_Right_Inches,
-                                          double Front_Left_Inches)
-    {
-        int newLeftBottomTarget;
-        int newRightBottomTarget;
-        int newRightTopTarget;
-        int newLeftTopTarget;
-
-        // Ensure that the opmode is still active
-        if (opModeIsActive()) {
-
-
-            // Determine new target position, and pass to motor controller
-            //newLeftBottomTarget = robot.Back_Left.getCurrentPosition() + (int)(Back_Left_Inches * COUNTS_PER_INCH);
-            //newRightBottomTarget = robot.Back_Right.getCurrentPosition() + (int)(Back_Right_Inches * (COUNTS_PER_INCH));
-            //newRightTopTarget = robot.Front_Right.getCurrentPosition() + (int) (Front_Right_Inches * COUNTS_PER_INCH);
-            newLeftTopTarget = robot.Front_Left.getCurrentPosition() + (int) (Front_Left_Inches * COUNTS_PER_INCH);
-
-            telemetry.addData("Current Position : %7d", robot.Front_Left.getCurrentPosition());
-            telemetry.addData("New Position : %7d", newLeftTopTarget);
-            telemetry.update();
-            sleep(2000);
-            //robot.Back_Left.setTargetPosition(newLeftBottomTarget);
-            //robot.Back_Right.setTargetPosition(newRightBottomTarget);
-            //robot.Front_Right.setTargetPosition(newRightTopTarget);
-            robot.Front_Left.setTargetPosition(newLeftTopTarget);
-
-            // Turn On RUN_TO_POSITION
-            //robot.Back_Left.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-            //robot.Back_Right.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-            robot.Front_Left.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-            //robot.Front_Right.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-            // reset the timeout time and start motion.
-            runtime.reset();
-            //robot.Back_Left.setPower(Math.abs(speed));
-            //robot.Back_Right.setPower(Math.abs(speed));
-            robot.Front_Left.setPower(Math.abs(speed));
-            //robot.Front_Right.setPower(Math.abs(speed));
-
-            // keep looping while we are still active, and there is time left, and both motors are running.
-            while (opModeIsActive() &&
-                    (robot.Front_Left.isBusy() )) {
-
-                // Display it for the driver.
-                //telemetry.addData("Path1",  "Running to %7d :%7d", newLeftBottomTarget,newRightBottomTarget,newLeftTopTarget,newRightTopTarget);
-                telemetry.addData("Path1",  "Running to %7d : ", newLeftTopTarget);
-                telemetry.addData("Path2",  "Running at %7d : ",
-                        robot.Front_Left.getCurrentPosition());
-                //robot.Front_Left.getCurrentPosition();
-                //robot.Front_Right.getCurrentPosition();
-                telemetry.update();
-            }
-
-            // Stop all motion;
-            //robot.Back_Left.setPower(0);
-            //robot.Back_Right.setPower(0);
-            robot.Front_Left.setPower(0);
-            //robot.Front_Right.setPower(0);
-
-            // Turn off RUN_TO_POSITION
-            robot.Back_Left.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            robot.Back_Right.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            robot.Front_Left.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            robot.Front_Right.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-            //  sleep(250);   // optional pause after each move
-
-        }
-
-    }
-
-    public void encoderDriveWithoutTimeFrontRight ( double speed,
-                                                   double Back_Left_Inches,
-                                                   double Back_Right_Inches,
-                                                   double Front_Right_Inches,
-                                                   double Front_Left_Inches)
-    {
-        int newLeftBottomTarget;
-        int newRightBottomTarget;
-        int newRightTopTarget;
-        int newLeftTopTarget;
-
-        // Ensure that the opmode is still active
-        if (opModeIsActive()) {
-
-
-            // Determine new target position, and pass to motor controller
-            //newLeftBottomTarget = robot.Back_Left.getCurrentPosition() + (int)(Back_Left_Inches * COUNTS_PER_INCH);
-            //newRightBottomTarget = robot.Back_Right.getCurrentPosition() + (int)(Back_Right_Inches * (COUNTS_PER_INCH));
-            newRightTopTarget = robot.Front_Right.getCurrentPosition() + (int) (Front_Right_Inches * COUNTS_PER_INCH);
-            //newLeftTopTarget = robot.Front_Left.getCurrentPosition() + (int) (Front_Left_Inches * COUNTS_PER_INCH);
-
-            telemetry.addData("Current Position : %7d", robot.Front_Right.getCurrentPosition());
-            telemetry.addData("New Position : %7d", newRightTopTarget);
-            telemetry.update();
-            sleep(2000);
-            //robot.Back_Left.setTargetPosition(newLeftBottomTarget);
-            //robot.Back_Right.setTargetPosition(newRightBottomTarget);
-            robot.Front_Right.setTargetPosition(newRightTopTarget);
-            //robot.Front_Left.setTargetPosition(newLeftTopTarget);
-
-            // Turn On RUN_TO_POSITION
-            //robot.Back_Left.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-            //robot.Back_Right.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-            //robot.Front_Left.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-            robot.Front_Right.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-            // reset the timeout time and start motion.
-            runtime.reset();
-            //robot.Back_Left.setPower(Math.abs(speed));
-            //robot.Back_Right.setPower(Math.abs(speed));
-            //robot.Front_Left.setPower(Math.abs(speed));
-            robot.Front_Right.setPower(Math.abs(speed));
-
-            // keep looping while we are still active, and there is time left, and both motors are running.
-            while (opModeIsActive() &&
-                    (robot.Front_Right.isBusy() )) {
-
-                // Display it for the driver.
-                //telemetry.addData("Path1",  "Running to %7d :%7d", newLeftBottomTarget,newRightBottomTarget,newLeftTopTarget,newRightTopTarget);
-                telemetry.addData("Path1",  "Running to %7d : ", newRightTopTarget);
-                telemetry.addData("Path2",  "Running at %7d : ",
-                        robot.Front_Right.getCurrentPosition());
-                //robot.Front_Left.getCurrentPosition();
-                //robot.Front_Right.getCurrentPosition();
-                telemetry.update();
-            }
-
-            // Stop all motion;
-            //robot.Back_Left.setPower(0);
-            //robot.Back_Right.setPower(0);
-            //robot.Front_Left.setPower(0);
             robot.Front_Right.setPower(0);
 
             // Turn off RUN_TO_POSITION
@@ -942,9 +913,9 @@ public class B1SpinPark extends LinearOpMode {
         /*
          * The core values which define the location and size of the sample regions
          */
-        static final Point REGION1_TOPLEFT_ANCHOR_POINT = new Point(75,58);
+        static final Point REGION1_TOPLEFT_ANCHOR_POINT = new Point(0,160);
 
-        static final int REGION_WIDTH = 140;
+        static final int REGION_WIDTH = 155;
         static final int REGION_HEIGHT = 66;
 
         final int THRESHOLD = 25;//150;
