+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Competition/R1ScanParkDelay.java
@@ -1,12 +1,11 @@
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Competition;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.disabled.Saketzanehardware;
 import org.opencv.core.Core;
 import org.opencv.core.Mat;
 import org.opencv.core.Point;
@@ -21,7 +20,7 @@ import org.openftc.easyopencv.OpenCvPipeline;
 
 
 @Autonomous
-public class R1SpinPark extends LinearOpMode {
+public class R1ScanParkDelay extends LinearOpMode {
 
     private static final long SLEEP_10 = 10;
     private static final long SLEEP_25 = 25;
@@ -30,13 +29,13 @@ public class R1SpinPark extends LinearOpMode {
 
     OpenCvInternalCamera phoneCam;
     SkystoneDeterminationPipeline pipeline;
-    Saketzanehardware robot = new Saketzanehardware();
+    zanehardware robot = new zanehardware();
     private ElapsedTime runtime = new ElapsedTime();
 
     static final double     COUNTS_PER_MOTOR_REV    = 537.6;//356.3 ;    // eg: DC Motor Encoder
     static final double     DRIVE_GEAR_REDUCTION    = 1.0 ;     // This is < 1.0 if geared UP
     static final double     WHEEL_DIAMETER_INCHES   = 3.77953 ;     // For figuring circumference
-    static final double     COUNTS_PER_INCH = 122.600924;
+    static final double     COUNTS_PER_INCH = 45;
             //first hundred digits of pi fr more accuracy
 
 
@@ -74,11 +73,16 @@ public class R1SpinPark extends LinearOpMode {
 
         robot.init(hardwareMap);
 
-        waitForStart();
 
+
+
+        waitForStart();
+        sleep(globalvar.delaytime);
 
         double FORWARD_SPEED = 0.5;
 
+
+        //scanShippingElementTest();
 
 
         while (opModeIsActive()) {
@@ -86,48 +90,50 @@ public class R1SpinPark extends LinearOpMode {
 
 
 
-            encoderDriveWithTimeLeft(.5,.5);
-            sleep(100);
-            encoderDriveWithTimeForward(2.7,1);
-            sleep(100);
-            spinRight(5, .3);
-            sleep(100);
-            //encoderDriveWithTimeLeft(.4,.5);
-            //sleep(100);
-            encoderDriveWithTimeBackward(1,1);
-            sleep(100);
-            encoderDriveWithTimeTurnLeft(1.5,.5);
-            sleep(100);
-            encoderDriveWithTimeForward(5.5,1);
-            stop();
 
-            double moveLength = 28;
+
+
+            encoderDriveWithoutTime(.5,1,1,1,1);
+            robot.Slider.setPower(1);
+            sleep(600);
+            robot.Slider.setPower(0.2);
+            sleep(200);
+            double moveLength = 33.5;
             int position = 1;
-            robotMoveToShippingElement(1);
-            sleep(5000);
+
+            //spinAndComeBack();
+            //stop();
+            robotMoveToShippingElement(13);
+
 
             boolean positionShippingElement = scanShippingElement();
             telemetry.addData("Is the shipping element present",positionShippingElement);
             telemetry.update();
-            sleep(1000);
+            sleep(100);
             if (positionShippingElement){
+                telemetry.addData("The position is",position);
+                telemetry.update();
                 moveToShippingHub(moveLength, position);
             }
 
             else {
 
-                moveToSecondDuck(6);
+                moveToSecondDuck(8);
                 positionShippingElement = scanShippingElement();
                 telemetry.addData("Is the shipping element present",positionShippingElement);
                 telemetry.update();
-                sleep(1000);
+                sleep(100);
                 if (positionShippingElement){
                     position = 2;
+                    telemetry.addData("The position is",position);
+                    telemetry.update();
 
-                    moveToShippingHub(moveLength - 6, position);
+                    moveToShippingHub(moveLength - 8, position);
                 }
                 else {
                     position = 3;
+                    telemetry.addData("The position is",position);
+                    telemetry.update();
 
                     moveToShippingHub(moveLength - 6, position);
                 }
@@ -156,11 +162,49 @@ public class R1SpinPark extends LinearOpMode {
         }
     }
 
+    private void spinAndComeBack() {
+
+        encoderDriveWithoutTime(.5,4,-4,4,-4);
+        sleep(100);
+        encoderDriveWithoutTime(.3,26,26,26,26);
+        sleep(100);
+        spinLeft(4,.4);
+        sleep(100);
+        encoderDriveWithoutTime(.5,-1,1,1,-1);
+        sleep(100);
+        encoderDriveWithoutTime(.5,6,-6,6,-6);
+        sleep(100);
+        encoderDriveWithoutTime(.5,-19,-19,-19,-19);
+        sleep(100);
+        turn90Left();
+    }
+
+    private void turn90Left() {
+
+        encoderDriveWithoutTime(.5,-19.5,19.5,19.5,-19.5);
+    }
+
+
+    private void turn90LeftMore() {
+
+        encoderDriveWithoutTime(.5,-20.5,20.5,20.5,-20.5);
+    }
+    private void turn90Right() {
+
+        encoderDriveWithoutTime(.5,18.5,-18.5,-18.5,18.5);
+    }
     private void moveToWarehouse() {
         //turn to the left for blue
-        encoderDriveWithoutTime(0.5,4,4,4,4);
+        //encoderDriveWithoutTime(0.5,4,4,4,4);
         //go forward
-        encoderDriveWithoutTime(0.5, 48,48,48,48);
+        //encoderDriveWithoutTime(0.5, 48,48,48,48);
+        encoderDriveWithoutTime(.8,5,5,5,5);
+        sleep(100);
+        turn90Left();
+        telemetry.addData("Before final move to warehouse","");
+        telemetry.update();
+        encoderDriveWithoutTime(.8,65,65,65,65);
+
     }
 
     private void placeFreightCorrectLocation(int position) {
@@ -168,13 +212,47 @@ public class R1SpinPark extends LinearOpMode {
         switch (position) {
             case 1:
                 //adjust arm position and drop cube
+                double len = 2;
+
+                robot.Sweeper.setPosition(0.4);
+                sleep(50);
+                len = 2;
+                encoderDriveWithoutTime(.5,-len,-len,-len,-len);
+                sleep(100);
                 break;
+
+
+
+
             case 2:
+                robot.Slider.setPower(0.7);
+                sleep(900);
+                robot.Slider.setPower(0.2);
+
+
+                robot.Sweeper.setPosition(0.4);
+                sleep(50);
+                len = 2;
+                encoderDriveWithoutTime(.5,-len,-len,-len,-len);
+                sleep(100);
                 //adjust arm position and drop cube
                 break;
             default:
+
+
+                robot.Slider.setPower(1);
+                sleep(1300);
+                robot.Slider.setPower(0.1);
+
+                robot.Sweeper.setPosition(0);
+                sleep(50);
+                len = 2;
+                encoderDriveWithoutTime(.5,-len,-len,-len,-len);
+                sleep(100);
                 //adjust arm position and drop cube
                 break;
+
+
 
         }
 
@@ -182,15 +260,25 @@ public class R1SpinPark extends LinearOpMode {
 
     private void moveToSecondDuck(double moveLength) {
 
-        //encoderDriveWithoutTime(0.3, moveLength, moveLength, moveLength, moveLength);
-        encoderDriveWithTime(0.3,1,1,1,1,2);
+        encoderDriveWithoutTime(0.3, -moveLength, moveLength, -moveLength, moveLength);
+        //encoderDriveWithTime(0.3,1,1,1,1,2);
     }
 
     private void moveToShippingHub(double moveLength, int position) {
 
-        encoderDriveWithoutTime(0.5, -moveLength, -moveLength, -moveLength, -moveLength);
-        encoderDriveWithoutTime(.3, -6,6,-6,6);
-        placeFreightCorrectLocation(position);
+        encoderDriveWithoutTime(0.5, -moveLength, moveLength, -moveLength, moveLength);
+        sleep(100);
+        turn90LeftMore();
+        telemetry.addData("Its working", "");
+        telemetry.update();
+        turn90LeftMore();
+        sleep(100);
+        //robot.Sweeper.setPosition(0);
+        encoderDriveWithoutTime(0.5,4,-4,4,-4);
+        encoderDriveWithoutTime(0.5, -10.5, -10.5, -10.5, -10.5);
+
+        //encoderDriveWithoutTime(.3, -6,6,-6,6);
+        //placeFreightCorrectLocation(position);
     }
 
     private boolean scanShippingElement() {
@@ -198,9 +286,40 @@ public class R1SpinPark extends LinearOpMode {
         sleep(1000);
         int avg1 = pipeline.getAnalysis();
 
+        final int THRESHOLD = 150;//150;
+        telemetry.addData("Scan Threshhold:",avg1);
+        telemetry.update();
+        sleep(1000);
+
+
+        if(avg1 > THRESHOLD){
+            return true;
+        }else{
+            return false;
+        }
+
+
+    }
+    private boolean scanShippingElementTest() {
+
+        while (opModeIsActive()) {
+            sleep(100);
+            int avg1 = pipeline.getAnalysis();
+
+            final int THRESHOLD = 25;//150;
+            telemetry.addData("Scan Threshhold:",avg1);
+            telemetry.update();
+            sleep(2000);
+
+
+        }
+        sleep(100);
+        int avg1 = pipeline.getAnalysis();
+
         final int THRESHOLD = 25;//150;
         telemetry.addData("Scan Threshhold:",avg1);
         telemetry.update();
+        sleep(5000);
 
 
         if(avg1 > THRESHOLD){
@@ -215,8 +334,8 @@ public class R1SpinPark extends LinearOpMode {
     private void robotMoveToShippingElement(double tmpmoveLength) {
         telemetry.addData("Going toward Shipping element", tmpmoveLength);
         telemetry.update();
-        //encoderDriveWithoutTime(0.3, -tmpmoveLength, -tmpmoveLength, -tmpmoveLength, -tmpmoveLength);
-        encoderDriveWithTime(1,-1,-1,-1,-1,10);
+        encoderDriveWithoutTime(0.3, tmpmoveLength, tmpmoveLength, tmpmoveLength, tmpmoveLength);
+        //encoderDriveWithTime(1,-1,-1,-1,-1,10);
     }
 
 
@@ -640,136 +759,6 @@ public class R1SpinPark extends LinearOpMode {
 
         }
     }
-
-    public void encoderDriveWithTimeTurnRight(
-            double timeoutS,
-            double speed) {
-
-
-        // Ensure that the opmode is still active
-        if (opModeIsActive()) {
-
-
-            // Determine new target position, and pass to motor controller
-            /*robot.Back_Right.setDirection(DcMotorSimple.Direction.FORWARD);
-            robot.Front_Left.setDirection(DcMotorSimple.Direction.REVERSE);
-            robot.Front_Right.setDirection(DcMotorSimple.Direction.FORWARD);
-            robot.Back_Left.setDirection(DcMotorSimple.Direction.REVERSE);*/
-
-            // Turn On RUN_TO_POSITION
-            //robot.Back_Left.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            //robot.Back_Right.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            robot.Front_Left.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            robot.Front_Right.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-
-            // reset the timeout time and start motion.
-            runtime.reset();
-            //robot.Back_Left.setPower(speed);
-            //robot.Back_Right.setPower(-speed);
-            robot.Front_Left.setPower(-speed);
-            robot.Front_Right.setPower(speed);
-
-
-            // keep looping while we are still active, and there is time left, and both motors are running.
-            while (opModeIsActive() &&
-                    (runtime.seconds() < timeoutS)) {
-                //&&
-                //(robot.Back_Left.isBusy() && robot.Back_Right.isBusy() && robot.Front_Left.isBusy() && robot.Front_Right.isBusy())) {
-
-                // Display it for the driver.
-                /*telemetry.addData("Path1",  "Running to %7d :%7d", newLeftBottomTarget,newRightBottomTarget,newLeftTopTarget,newRightTopTarget);
-                telemetry.addData("Path2",  "Running at %7d :%7d",
-                        robot.Back_Left.getCurrentPosition(),
-                        robot.Back_Right.getCurrentPosition());
-                robot.Front_Left.getCurrentPosition();
-                robot.Front_Right.getCurrentPosition();*/
-                telemetry.addData("Going Right for ", timeoutS - runtime.seconds());
-
-
-                telemetry.update();
-            }
-
-            // Stop all motion;
-            robot.Back_Left.setPower(0);
-            robot.Back_Right.setPower(0);
-            robot.Front_Left.setPower(0);
-            robot.Front_Right.setPower(0);
-
-            // Turn off RUN_TO_POSITION
-            robot.Back_Left.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            robot.Back_Right.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            robot.Front_Left.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            robot.Front_Right.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-
-            //  sleep(250);   // optional pause after each move
-
-        }
-    }
-
-    public void encoderDriveWithTimeTurnLeft(
-            double timeoutS,
-            double speed) {
-
-
-        // Ensure that the opmode is still active
-        if (opModeIsActive()) {
-
-
-            // Determine new target position, and pass to motor controller
-            /*robot.Back_Right.setDirection(DcMotorSimple.Direction.REVERSE);
-            robot.Front_Left.setDirection(DcMotorSimple.Direction.FORWARD);
-            robot.Front_Right.setDirection(DcMotorSimple.Direction.REVERSE);
-            robot.Back_Left.setDirection(DcMotorSimple.Direction.FORWARD);*/
-
-            // Turn On RUN_TO_POSITION
-            robot.Back_Left.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            robot.Back_Right.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            //robot.Front_Left.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            //robot.Front_Right.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-
-            // reset the timeout time and start motion.
-            runtime.reset();
-            robot.Back_Left.setPower(-speed);
-            robot.Back_Right.setPower(speed);
-            //robot.Front_Left.setPower(speed);
-            //robot.Front_Right.setPower(-speed);
-
-            // keep looping while we are still active, and there is time left, and both motors are running.
-            while (opModeIsActive() &&
-                    (runtime.seconds() < timeoutS)) {
-                //&&
-                //(robot.Back_Left.isBusy() && robot.Back_Right.isBusy() && robot.Front_Left.isBusy() && robot.Front_Right.isBusy())) {
-
-                // Display it for the driver.
-                /*telemetry.addData("Path1",  "Running to %7d :%7d", newLeftBottomTarget,newRightBottomTarget,newLeftTopTarget,newRightTopTarget);
-                telemetry.addData("Path2",  "Running at %7d :%7d",
-                        robot.Back_Left.getCurrentPosition(),
-                        robot.Back_Right.getCurrentPosition());
-                robot.Front_Left.getCurrentPosition();
-                robot.Front_Right.getCurrentPosition();*/
-                telemetry.addData("Turning Left for ", timeoutS - runtime.seconds());
-
-
-                telemetry.update();
-            }
-
-            // Stop all motion;
-            robot.Back_Left.setPower(0);
-            robot.Back_Right.setPower(0);
-            robot.Front_Left.setPower(0);
-            robot.Front_Right.setPower(0);
-
-            // Turn off RUN_TO_POSITION
-            robot.Back_Left.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            robot.Back_Right.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            robot.Front_Left.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            robot.Front_Right.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-
-            //  sleep(250);   // optional pause after each move
-
-        }
-    }
-
     public void spinLeft(
             double timeoutS,
             double speed) {
@@ -906,9 +895,9 @@ public class R1SpinPark extends LinearOpMode {
         /*
          * The core values which define the location and size of the sample regions
          */
-        static final Point REGION1_TOPLEFT_ANCHOR_POINT = new Point(75,58);
+        static final Point REGION1_TOPLEFT_ANCHOR_POINT = new Point(0,160);
 
-        static final int REGION_WIDTH = 140;
+        static final int REGION_WIDTH = 155;
         static final int REGION_HEIGHT = 66;
 
         final int THRESHOLD = 25;//150;
