+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Competition/globalvar.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Competition;
 
 public class globalvar {
     static int delaytime=10000;
