+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Competition/zanehardware.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Competition;
 
 
 import com.qualcomm.robotcore.hardware.CRServo;
