+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/disabled/AutonomousFreightFrenzyTEST.java
@@ -6,7 +6,7 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.zanehardware;
+import org.firstinspires.ftc.teamcode.Competition.zanehardware;
 import org.opencv.core.Core;
 import org.opencv.core.Mat;
 import org.opencv.core.Point;
