+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/disabled/JosephIntakeTest.java
@@ -1,15 +1,10 @@
 package org.firstinspires.ftc.teamcode.disabled;
 
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
-import org.firstinspires.ftc.teamcode.zanehardware;
+import org.firstinspires.ftc.teamcode.Competition.zanehardware;
 
 @TeleOp(name = "Joesph intake test")
 public class JosephIntakeTest extends LinearOpMode {
