+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/disabled/victorB1automous.java
@@ -1,16 +1,15 @@
 package org.firstinspires.ftc.teamcode.disabled;
 
-import static org.firstinspires.ftc.teamcode.zanehardware.COUNTS_PER_INCH;
+import static org.firstinspires.ftc.teamcode.Competition.zanehardware.COUNTS_PER_INCH;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
-import org.firstinspires.ftc.teamcode.zanehardware;
+import org.firstinspires.ftc.teamcode.Competition.zanehardware;
 
 @Disabled
 @Autonomous(name = "Victors B1 automous")
